+++ src/Products/PluginIndexes/UUIDIndex/UUIDIndex.py
@@ -43,7 +43,7 @@ class UUIDIndex(UnIndex):
         {'label': 'Browse', 'action': 'manage_browse'},
     )
 
-    query_options = ['query', 'range']
+    query_options = ['query', 'range', 'not']
 
     manage = manage_main = DTMLFile('dtml/manageUUIDIndex', globals())
     manage_main._setName('manage_main')